+++ tests/framework_tornado_r3/test_async_application.py
@@ -679,6 +679,34 @@ class AllTests(object):
         self.assertEqual(response.body, external.RESPONSE)
 
     # The port number 8989 matches the port number in MockExternalHTTPServer
+    # The count matches the number of requests handler executes after inital
+    # get request.
+    scoped_metrics = [
+            ('Function/_test_async_application:'
+             'AsyncExternalCountHandler.get', 1),
+            ('External/localhost:8989/tornado.httpclient/', 2),
+    ]
+    rollup_metrics = [('External/allWeb', 2), ('External/all', 2)]
+
+    # Test to make sure multiple concurrent requests are handled correctly.
+    @tornado_validate_transaction_cache_empty()
+    @tornado_validate_errors()
+    @tornado_validate_count_transaction_metrics(
+            '_test_async_application:AsyncExternalCountHandler.get',
+            scoped_metrics=scoped_metrics,
+            rollup_metrics=rollup_metrics,
+    )
+    def test_multiple_requests(self):
+        port = 8989
+        num_requests = 2
+
+        with MockExternalHTTPServer() as external:
+            response = self.fetch_response('/async-client/%s/%s' % (port,
+                    num_requests))
+
+        assert response.code == 200
+
+    # The port number 8989 matches the port number in MockExternalHTTPServer
     scoped_metrics = [('Function/_test_async_application:'
         'CustomImplCurlAsyncRequestHandler.get', 1),
         ('Function/_test_async_application:CustomImplCurlAsyncRequestHandler.'